# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ADEX documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "ADEX API Server"
        }
    ],
    "paths": {
...
        "/task": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Создать новую задачу",
                "description": "Создать новую задачу (только для пользователей с типом заказчик)",
                "operationId": "createTask",
                "requestBody": {
                    "description": "Заполните поля для создания задачи",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "categories"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Название задачи"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "1200"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание задачи"
                                    },
                                    "categories": {
                                        "type": "integer",
                                        "example": [
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Создать рекламу"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Создать рекламу на тему пиар"
                                        },
                                        "price": {
                                            "type": "integer",
                                            "example": "1200"
                                        },
                                        "views": {
                                            "type": "integer",
                                            "example": "15"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Busy"
                                        },
                                        "categories": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Video"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-04-04"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unautorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Unautorized"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Получить задачи с фильтрами",
                "operationId": "getTasks",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Создать рекламу"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Создать рекламу на тему пиар"
                                        },
                                        "price": {
                                            "type": "integer",
                                            "example": "1200"
                                        },
                                        "views": {
                                            "type": "integer",
                                            "example": "15"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Busy"
                                        },
                                        "categories": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Video"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-04-04"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/task/{taskId}/status": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Изменить статус задачи",
                "description": "Изменить статус задачи ( доступно только для пользователя с типом заказчик)",
                "operationId": "setTaskStatus",
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Task not found"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unautorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Unautorized"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    }
}